requirements.txt
asgiref
beautifulsoup4
certifi
cffi
chardet
cryptography
defusedxml
Django
django-braces
django-cors-headers
django-filter
django-oauth-toolkit
django-redis
django-rest-framework-social-oauth2
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
future
idna
importlib-metadata
line-bot-sdk
Markdown
oauthlib
pycparser
PyJWT
python3-openid
pytz
redis
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
soupsieve
sqlparse
urllib3
wget
zipp
